+++ berkeley-function-call-leaderboard/bfcl/eval_checker/model_metadata.py
@@ -1,7 +1,7 @@
 from bfcl.model_handler.handler_map import local_inference_handler_map
 
 MODEL_METADATA_MAPPING = {
-    "palmyra-x-004-FC": [
+    "palmyra-x-004": [
         "palmyra-x-004 (FC)",
         "https://writer.com/engineering/actions-with-palmyra-x-004/",
         "Writer",
@@ -771,7 +771,7 @@ INPUT_PRICE_PER_MILLION_TOKEN = {
     "command-r-plus-FC-optimized": 3,
     "command-r-plus-optimized": 3,
     "yi-large-fc": 3,
-    "palmyra-x-004-FC": 5,
+    "palmyra-x-004": 5,
 }
 
 OUTPUT_PRICE_PER_MILLION_TOKEN = {
@@ -834,7 +834,7 @@ OUTPUT_PRICE_PER_MILLION_TOKEN = {
     "command-r-plus-FC-optimized": 15,
     "command-r-plus-optimized": 15,
     "yi-large-fc": 3,
-    "palmyra-x-004-FC": 12,
+    "palmyra-x-004": 12,
 }
 
 # The latency of the open-source models are hardcoded here.