+++ berkeley-function-call-leaderboard/bfcl/model_handler/proprietary_model/writer.py
@@ -10,6 +10,7 @@ class WriterHandler(OpenAIHandler):
         super().__init__(model_name, temperature)
         self.model_style = ModelStyle.WRITER
         self.client = Writer(api_key=os.getenv("WRITER_API_KEY"))
+        self.is_fc_model = True
 
     #### FC methods ####